+++ internal/timespec/timespec.go
@@ -1,4 +1,4 @@
-package cephfs
+package timespec
 
 /*
 #include <time.h>
@@ -6,6 +6,8 @@ package cephfs
 import "C"
 
 import (
+	"unsafe"
+
 	"golang.org/x/sys/unix"
 )
 
@@ -14,7 +16,13 @@ import (
 // apis that could be lossy with the use of Go time types.
 type Timespec unix.Timespec
 
-func cStructToTimespec(t C.struct_timespec) Timespec {
+// CTimespecPtr is an unsafe pointer wrapping C's `struct timespec`.
+type CTimespecPtr unsafe.Pointer
+
+// CStructToTimespec creates a new Timespec for the C 'struct timespec'.
+func CStructToTimespec(cts CTimespecPtr) Timespec {
+	t := (*C.struct_timespec)(cts)
+
 	return Timespec{
 		Sec:  int64(t.tv_sec),
 		Nsec: int64(t.tv_nsec),
